flask_sell_me/requirements.txt
blinker==1.8.2
click==8.1.7
filelock==3.14.0
Flask==3.0.3
fsspec==2024.5.0
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==2.1.5
mpmath==1.3.0
networkx==3.3
numpy==1.26.4
opencv-python==4.9.0.80
pillow==10.3.0
sympy==1.12
torch==2.2.2
torchvision==0.17.2
typing_extensions==4.11.0
Werkzeug==3.0.3
